Add a basic broker implementation
This is the minimal config and code required to implement the service
broker interface and compile a broker server. No operations are
currently supported.

We choose to import the broker as a library. This is because a lot of
the implementation is the same across brokers, and therefore can be
shared.

(Thanks @henrytk :smile:)